scraper/requirements.txt
adal==1.2.1
aenum==2.1.2
alembic==1.0.11
amqp==2.5.0
appdirs==1.4.3
argh==0.26.2
arrow==0.15.2
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.2.1
astroquery==0.3.9
async-timeout==3.0.1
asyncio-nats-client==0.9.2
attrdict==2.0.1
attrs==19.1.0
azure-common==1.1.23
azure-servicebus==0.50.1
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
billiard==3.6.0.0
bleach==3.1.0
blessings==1.7
blinker==1.4
boto==2.49.0
bpython==0.17.1
bunq-sdk==1.0.0
cached-property==1.5.1
cachetools==3.1.1
cairocffi==1.0.2
CairoSVG==2.4.0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cmake-example==0.0.1
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
cssselect2==0.2.1
curtsies==0.3.0
cycler==0.10.0
Cython==0.29.16
dataclasses==0.7
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.2
django-celery-results==1.1.2
django-colorfield==0.1.14
django-cors-headers==2.2.0
django-debug-toolbar==2.0
django-extensions==2.1.6
django-extra-fields==1.0.0
django-filter==1.1.0
django-money==0.13.2
django-oauth-toolkit==1.2.0
django-polls==0.1
django-redis==4.9.0
django-registration-redux==2.4
django-render-block==0.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-templated-email==2.2.0
django-tracer==0.9.3
djangorestframework==3.9.4
docutils==0.15.2
drf-extensions==0.3.1
drf-writable-nested==0.4.2
drf-yasg==1.16.0
elementpath==1.1.8
entrypoints==0.3
esper==1.2
example==0.1.0
fasttext==0.9.2
filelock==3.0.12
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-mongoengine==0.9.5
Flask-SQLAlchemy==2.4.0
Flask-User==1.0.1.5
Flask-WTF==0.14.2
flower==0.9.3
future==0.18.2
gensim==3.8.1
geonamescache==0.21
geoplotlib==0.3.2
google-auth==1.6.3
googleads==19.0.1
greenlet==0.4.15
html5lib==1.0.1
httpretty==0.9.6
idna==2.8
imutils==0.5.2
inflection==0.3.1
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
isort==4.3.20
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.13.3
jeepney==0.4
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.0
jplephem==2.9
jsonify==0.5
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
keyring==19.0.2
kiwisolver==1.1.0
kombu==4.6.3
lazy-object-proxy==1.4.1
llvmlite==0.29.0
lorem==0.1.1
lxml==4.3.3
Mako==1.0.14
marble==0.0.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mongoengine==0.18.2
msrest==0.6.8
msrestazure==0.6.1
Naked==0.1.31
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.4
nlpaug==0.0.14
nltk==3.4.5
nose2==0.9.1
notebook==5.7.8
numba==0.44.0
numpy==1.17.4
oauthlib==3.0.1
openapi-codec==1.3.2
opencv-python==4.1.0.25
OrbitalPy==0.7.0
pandas==1.0.3
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.4.0
passlib==1.7.1
pathtools==0.1.2
pdfrw==0.4
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
pika==0.11.2
plotly==3.10.0
poliastro==0.12.0
proj==0.1.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.2
ptyprocess==0.6.0
py-moneyed==0.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pybind11==2.5.0
pycodestyle==2.5.0
pycparser==2.19
pycryptodomex==3.8.2
pydotplus==2.0.2
pygame==1.9.6
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.8.0
PyNaCl==1.3.0
pyOpenSSL==18.0.0
pyparsing==2.4.2
Pyphen==0.9.5
pyrsistent==0.15.2
pyspellchecker==0.5.4
python-dateutil==2.8.1
python-editor==1.0.4
python-example==0.0.1
python-json-logger==0.1.9
python-status==1.0.1
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.2
pyzbar==0.1.8
pyzmq==18.0.1
qrcode==6.1
qtconsole==4.5.1
redis==3.2.1
regex==2020.5.14
Represent==1.6.0
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
retrying==1.3.3
rope==0.14.0
rq==1.0
rsa==4.0
ruamel.yaml==0.15.97
sacremoses==0.0.43
scipy==1.3.0
seaborn==0.9.0
SecretStorage==3.1.1
selenium==3.141.0
Send2Trash==1.5.0
sentencepiece==0.1.91
sepaxml==2.1.1
sgp4==1.4
simplejson==3.16.0
simulator==0.0.1
six==1.13.0
smart-open==1.9.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.1
SQLAlchemy==1.3.6
sqlparse==0.3.0
status==0.2.5
suds-jurko==0.6
tailer==0.4.1
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
timestring==1.6.4
tinycss2==1.0.2
tokenizers==0.7.0
torch==1.5.0
tornado==6.0.2
tqdm==4.46.0
traitlets==4.3.2
transformers==2.10.0
typed-ast==1.4.0
typing==3.7.4
uamqp==1.2.1
uritemplate==3.0.0
urllib3==1.25.7
vine==1.3.0
virtualenv==16.6.1
watchdog==0.9.0
wcwidth==0.1.7
WeasyPrint==0.42.3
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.4.2
wrapt==1.11.1
WTForms==2.2.1
XlsxWriter==1.1.5
xmlschema==1.0.13
xmltodict==0.12.0
zeep==3.4.0
